Implement navigation and editing in view component
Added navigation to the editor component on edit click in the view component and implemented an employee editing form in the editor component. Also added routing for navigation to the editor component for an employee identified by id and fetched the relevant employee data on initialization of the editor component.