Add edit functionality to qualifications component
This update includes several significant changes to the qualifications component. An "edit qualification" button has been added to the qualifications list, enabling inline editing of existing qualifications. This was achieved by consolidating "add" and "edit" functionalities into a singular "openDialog" method, and renaming "AddQualificationDialogComponent" to "AddEditQualificationDialogComponent" to reflect this change. Also, a "validateQualification" method has been introduced to the qualifications service for validating the existence of a qualification.