Refactor skills section in editor component
The skills section in the editor component has been refactored for better usability and design. It now showcases skills in an expandable form rather than a checklist, utilizing MatIconModule, MatAccordion and MatExpansionPanel components from Angular Material. This change also saw the removal of 'checked' field in the qualifications mapping within the component and updating of 'ngModel' in the HTML from 'check' to 'skill'.