Refactor editor component to use forkJoin
This commit refactors the editor component to use 'forkJoin' from RxJS for simultaneous handling of qualifications and employees. The changes modify how the form for qualifications is populated and the way an employee is fetched and processed. The editor.component.html file is also updated for tracking qualifications by index instead of id.