Replace custom cookie consent dialog with ngx-cookieconsent
Replaced the custom implementation of the cookie consent dialog with a widely used third-party library ngx-cookieconsent for cleaner and more efficient cookie policy management. This change also includes deletion of custom cookie dialog code and configuration of the new cookie consent in the application config. Make sure to replace placeholder URLs with the actual policy URLs in the future.