Update Github Actions in CI pipeline
The Github Actions used in the CI pipeline configuration file have been updated. Specifically, the "checkout" and "cache" actions were upgraded from v2 to v4 and v1 to v4 respectively. These updates ensure that we're using the latest versions of the specified actions, improving the reliability and efficiency of our CI pipeline.